refactor file structures
