+++ op-conductor/conductor/service.go
@@ -292,7 +292,7 @@ func (oc *OpConductor) initRPCServer(ctx context.Context) error {
 	return nil
 }
 
-// init flashblocks handler
+// initFlashblocksHandler initializes the flashblocks handler
 func (c *OpConductor) initFlashblocksHandler(ctx context.Context) error {
 	if c.cfg.RollupBoostWsURL == "" && c.cfg.WebsocketServerPort <= 0 {
 		c.log.Info("flashblocks handler disabled, no rollup boost URL or websocket server port configured")
